% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums_by_band.R
\name{albums_by_bands}
\alias{albums_by_bands}
\title{Retrieve albums by a specific band}
\usage{
albums_by_bands(band.var)
}
\arguments{
\item{band.var}{The name of the band.}
}
\value{
data of Album, Year, and Rating for the specified band.
}
\description{
Filters the album for given artist, returns album details ordered by year.
}
\examples{
albums_by_bands("Beatles")
albums_by_bands("Nirvana")
}
