% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_bands.R
\name{band_album_comparison_chart}
\alias{band_album_comparison_chart}
\title{Creates a chart comparing album ratings of two bands or artists}
\usage{
band_album_comparison_chart(var.artist1, var.artist2)
}
\arguments{
\item{var.artist1}{The name of the first band or artist.}

\item{var.artist2}{The name of the second band or artist.}
}
\value{
A ggplot object that is a line chart. Artist 1 is red, Artist 2 is blue.
}
\description{
Creates a chart showing the album ratings for two specified artists,
on the x-axis rating and y-axis is the year.
}
\examples{
band_album_comparison_chart("Beatles", "Nirvana")
band_album_comparison_chart("Maroon 5", "Bruno Mars")
}
