% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums_by_band.R
\name{band_album_count}
\alias{band_album_count}
\title{Count the number of albums for a specific band}
\usage{
band_album_count(band.var)
}
\arguments{
\item{band.var}{The name of the band}
}
\value{
Prints the number albums to the console.
}
\description{
Filters albums for given artist, counts albums.
It then prints the count to the console.
}
\examples{
band_album_count("Beatles")
band_album_count("Nirvana")
}
