% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums_by_band.R
\name{band_mean_rating}
\alias{band_mean_rating}
\title{Calculate the mean album rating for a specific band/artist}
\usage{
band_mean_rating(band.var)
}
\arguments{
\item{band.var}{The name of the band.}
}
\value{
Prints the average rating to the console as a string, formatted to two decimal places.
        For example: "Average Rating: 10".
}
\description{
Filters albums by bands and average of their album ratings.
It then prints this average to the console.
}
\examples{
band_mean_rating("Beatles")
band_mean_rating("Nirvana")
}
