% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinyl.R
\name{missing_vinyl}
\alias{missing_vinyl}
\title{Find albums not owned on vinyl}
\usage{
missing_vinyl()
}
\value{
data with Album, Artist, and Rating for top-rated albums not owned on vinyl.
}
\description{
Find albums with a rating of 9 or higher
where the 'Vinyl' column is empty meaning that it's not owned.
results ordered by rating in descending order.
}
\examples{
missing_vinyl()
}
