% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_one_albums.R
\name{number_one_album}
\alias{number_one_album}
\title{List number one ranked albums given a specified year range}
\usage{
number_one_album(var.startyear, var.endyear)
}
\arguments{
\item{var.startyear}{The starting year.}

\item{var.endyear}{The ending year.}
}
\value{
data with Year, Album, and Artist for albums
        ranked number 1 in the specified range.
}
\description{
Find albums ranked as number 1 within a given start and end year inclusivly.
}
\examples{
number_one_album(2003, 2005)
number_one_album(1999, 1999)
}
