% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums_by_year.R
\name{year_albums}
\alias{year_albums}
\title{List top albums for a specific year}
\usage{
year_albums(year.var)
}
\arguments{
\item{year.var}{The year to filter albums by.}
}
\value{
data with Ranking, Album, and Artist for the specified year.
}
\description{
Filters the album for a given year, returns a selection of album details, ordered by their ranking.
}
\examples{
year_albums(2003)
year_albums(2000)
}
