% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vinyl.R
\name{year_most_vinyl}
\alias{year_most_vinyl}
\title{Find years with the most vinyl owned}
\usage{
year_most_vinyl()
}
\value{
data with Year and the count of vinly owned per year,
        sorted by n which is the count of vinyl owned for that year
}
\description{
Find album data for entries where 'Vinyl' is marked as "v" meaning it's owned
counts the number of owned vinyl albums per year
results ordered by descending order and based on the count.
}
\examples{
year_most_vinyl()
}
